#13 - Current state command and refactor (replace mixin with base block)
